Refactor load plugins modules
