+++ packages/gatsby/src/bootstrap/load-plugins/resolve-plugin.ts
@@ -1,0 +1,100 @@
+import path from "path"
+import fs from "fs"
+import { slash, createRequireFromPath } from "gatsby-core-utils"
+import { sync as existsSync } from "fs-exists-cached"
+import { warnOnIncompatiblePeerDependency } from "./validate"
+import { PackageJson } from "../../.."
+import { IPluginInfo, PluginRef } from "./types"
+import { createPluginId } from "./utils/create-id"
+import { createFileContentHash } from "./utils/create-hash"
+import reporter from "gatsby-cli/lib/reporter"
+import { isString } from "lodash"
+
+/**
+ * @param plugin
+ * This should be a plugin spec object where possible but can also be the
+ * name of a plugin.
+ *
+ * When it is a name, it can be a name of a local plugin, the name of a plugin
+ * located in node_modules, or a Gatsby internal plugin. In the last case the
+ * plugin will be an absolute path.
+ * @param rootDir
+ * This is the project location, from which are found the plugins
+ */
+export function resolvePlugin(plugin: PluginRef, rootDir: string): IPluginInfo {
+  const pluginName = isString(plugin) ? plugin : plugin.resolve
+
+  // Only find plugins when we're not given an absolute path
+  if (!existsSync(pluginName) && rootDir) {
+    // Find the plugin in the local plugins folder
+    const resolvedPath = slash(path.join(rootDir, `plugins/${pluginName}`))
+
+    if (existsSync(resolvedPath)) {
+      if (existsSync(`${resolvedPath}/package.json`)) {
+        const packageJSON = JSON.parse(
+          fs.readFileSync(`${resolvedPath}/package.json`, `utf-8`)
+        ) as PackageJson
+        const name = packageJSON.name || pluginName
+        warnOnIncompatiblePeerDependency(name, packageJSON)
+
+        return {
+          resolve: resolvedPath,
+          name,
+          id: createPluginId(name),
+          version:
+            packageJSON.version || createFileContentHash(resolvedPath, `**`),
+        }
+      } else {
+        // Make package.json a requirement for local plugins too
+        throw new Error(`Plugin ${pluginName} requires a package.json file`)
+      }
+    }
+  }
+
+  /**
+   * Here we have an absolute path to an internal plugin, or a name of a module
+   * which should be located in node_modules.
+   */
+  try {
+    const requireSource =
+      rootDir !== null
+        ? createRequireFromPath(`${rootDir}/:internal:`)
+        : require
+
+    // If the path is absolute, resolve the directory of the internal plugin,
+    // otherwise resolve the directory containing the package.json
+    const resolvedPath = slash(
+      path.dirname(
+        requireSource.resolve(
+          path.isAbsolute(pluginName)
+            ? pluginName
+            : `${pluginName}/package.json`
+        )
+      )
+    )
+
+    const packageJSON = JSON.parse(
+      fs.readFileSync(`${resolvedPath}/package.json`, `utf-8`)
+    )
+    warnOnIncompatiblePeerDependency(packageJSON.name, packageJSON)
+
+    return {
+      resolve: resolvedPath,
+      id: createPluginId(packageJSON.name),
+      name: packageJSON.name,
+      version: packageJSON.version,
+    }
+  } catch (err) {
+    if (process.env.gatsby_log_level === `verbose`) {
+      reporter.panicOnBuild(
+        `plugin "${pluginName} threw the following error:\n`,
+        err
+      )
+    } else {
+      reporter.panicOnBuild(
+        `There was a problem loading plugin "${pluginName}". Perhaps you need to install its package?\nUse --verbose to see actual error.`
+      )
+    }
+    throw new Error(`unreachable`)
+  }
+}
